Split up EntityShareModal and create EntityShareSettings component
