+++ graylog2-web-interface/src/components/permissions/EntityShareModal.jsx
@@ -2,19 +2,13 @@
 import * as React from 'react';
 import { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
 
-import type { GRN } from 'logic/permissions/types';
 import { useStore } from 'stores/connect';
 import { Spinner } from 'components/common';
 import { EntityShareStore, EntityShareActions } from 'stores/permissions/EntityShareStore';
-import EntityShareState from 'logic/permissions/EntityShareState';
 import BootstrapModalConfirm from 'components/bootstrap/BootstrapModalConfirm';
 
-import GranteesSelector, { type SelectionRequest } from './GranteesSelector';
-import GranteesList from './GranteesList';
-import DependenciesWarning from './DependenciesWarning';
-import ShareableEnityURL from './ShareableEnityURL';
+import EntityShareSettings from './EntityShareSettings';
 
 type Props = {
   description: string,
@@ -22,106 +16,6 @@ type Props = {
   entityTitle: string,
   entityType: string,
   onClose: () => void,
-};
-
-type ModalContentProps = {
-  description: $PropertyType<Props, 'description'>,
-  entityGRN: GRN,
-  entityShareState: EntityShareState,
-  setDisableSubmit: (boolean) => void,
-};
-
-const Section = styled.div`
-  margin-bottom: 25px;
-  :last-child {
-    margin-bottom: 0;
-  }
-`;
-
-const GranteesSelectorHeadline = styled.h5`
-  margin-bottom: 10px;
-`;
-
-const _filterAvailableGrantees = (availableGrantees, selectedGranteeRoles) => {
-  const availableGranteeRolesUserIds = selectedGranteeRoles.entrySeq().map(([granteeGRN]) => granteeGRN);
-
-  return availableGrantees.filter((grantee) => !availableGranteeRolesUserIds.includes(grantee.id));
-};
-
-const ModalContent = ({
-  entityShareState: {
-    activeShares,
-    availableGrantees,
-    availableRoles,
-    missingDependencies,
-    selectedGranteeRoles,
-    selectedGrantees,
-  },
-  description,
-  entityGRN,
-  setDisableSubmit,
-}: ModalContentProps) => {
-  const filteredGrantees = _filterAvailableGrantees(availableGrantees, selectedGranteeRoles);
-
-  const _handleSelection = ({ granteeId, roleId }: SelectionRequest) => {
-    setDisableSubmit(true);
-
-    return EntityShareActions.prepare(entityGRN, {
-      selected_grantee_roles: selectedGranteeRoles.merge({ [granteeId]: roleId }),
-    }).then((response) => {
-      setDisableSubmit(false);
-
-      return response;
-    });
-  };
-
-  const _handleDeletion = (granteeId: GRN) => {
-    setDisableSubmit(true);
-
-    return EntityShareActions.prepare(entityGRN, {
-      selected_grantee_roles: selectedGranteeRoles.remove(granteeId),
-    }).then((response) => {
-      setDisableSubmit(false);
-
-      return response;
-    });
-  };
-
-  return (
-    <>
-      <Section>
-        <GranteesSelectorHeadline>
-          Add collaborator
-        </GranteesSelectorHeadline>
-        <p>
-          {description}
-        </p>
-      </Section>
-      <Section>
-        <GranteesSelector availableGrantees={filteredGrantees}
-                          availableRoles={availableRoles}
-                          onSubmit={_handleSelection} />
-      </Section>
-      <Section>
-        <GranteesList activeShares={activeShares}
-                      availableRoles={availableRoles}
-                      entityGRN={entityGRN}
-                      onDelete={_handleDeletion}
-                      onRoleChange={_handleSelection}
-                      selectedGrantees={selectedGrantees}
-                      title="Current collaborators" />
-      </Section>
-      {missingDependencies && (
-        <Section>
-          <DependenciesWarning missingDependencies={missingDependencies}
-                               selectedGrantees={selectedGrantees} />
-        </Section>
-      )}
-      <Section>
-        <ShareableEnityURL />
-      </Section>
-    </>
-  );
 };
 
 const _generateGRN = (id, type) => `grn::::${type}:${id}`;
@@ -152,10 +46,10 @@ const EntityShareModal = ({ description, entityId, entityType, entityTitle, onCl
                            title={<>Sharing: {entityType}: <i>{entityTitle}</i></>}>
       <>
         {entityShareState ? (
-          <ModalContent description={description}
-                        entityGRN={entityGRN}
-                        entityShareState={entityShareState}
-                        setDisableSubmit={setDisableSubmit} />
+          <EntityShareSettings description={description}
+                               entityGRN={entityGRN}
+                               entityShareState={entityShareState}
+                               setDisableSubmit={setDisableSubmit} />
         ) : (
           <Spinner />
         )}
